change roullete function to receive a population
Before this commit, roullete accepted a RangeDictionary whose keys
represented the probability of each member being selected. This
puts a burden on the caller - convert a population to a
RangeDict. Moreover, it is expected that only the rou-
llete function will ever need this kind of thing.

Given all of that, I find that it makes for a cleaner architecture
to have `roullete` - and all the future population selectors -
receive a *population* and make the necessary transformations
they need.