make GASolver implement the Iterable protocol
This commit adds __iter__ and __next__ so that we can iterate
over a GASolver instance. It allows for a more pythonic
usage of the class. This commit also adds a new max
steps attribute that may be used to limit how
far the solver goes trying to find a
solution.